+++ alphabase/psm_reader/psm_reader.py
@@ -1,5 +1,6 @@
 import os
 import copy
+import io
 import pandas as pd
 import numpy as np
 
@@ -9,6 +10,7 @@ from alphabase.peptide.precursor import (
 )
 from alphabase.constants._const import CONST_FILE_FOLDER
 
+from alphabase.utils import get_delimiter
 from alphabase.yaml_utils import load_yaml
 
 def translate_other_modification(
@@ -166,6 +168,7 @@ class PSMReaderBase(object):
         self.keep_decoy = keep_decoy
         self._min_max_rt_norm = False
         self._engine_rt_unit = rt_unit
+        self._min_irt_value = -100
 
     @property
     def psm_df(self)->pd.DataFrame:
@@ -316,6 +319,9 @@ class PSMReaderBase(object):
         # to -log(evalue), as score is the larger the better
         pass
 
+    def _get_table_delimiter(self, _filename):
+        return get_delimiter(_filename)
+
     def normalize_rt(self):
         if 'rt' in self.psm_df.columns:
             if self._engine_rt_unit == 'second':
@@ -326,6 +332,12 @@ class PSMReaderBase(object):
             min_rt = self.psm_df.rt.min()
             if not self._min_max_rt_norm or min_rt > 0:
                 min_rt = 0
+            elif min_rt < self._min_irt_value: # iRT
+                self.psm_df.rt.values[
+                    self.psm_df.rt.values<self._min_irt_value
+                ] = self._min_irt_value
+                min_rt = self._min_irt_value
+            
             self.psm_df['rt_norm'] = (
                 self.psm_df.rt - min_rt
             ) / (self.psm_df.rt.max()-min_rt)
@@ -369,6 +381,19 @@ class PSMReaderBase(object):
             f'"{self.__class__}" must implement "_load_file()"'
         )
 
+    def _find_mapped_columns(self, origin_df:pd.DataFrame):
+        mapped_columns = {}
+        for col, map_col in self.column_mapping.items():
+            if isinstance(map_col, str):
+                if map_col in origin_df.columns:
+                    mapped_columns[col] = map_col
+            elif isinstance(map_col, (list,tuple)):
+                for other_col in map_col:
+                    if other_col in origin_df.columns:
+                        mapped_columns[col] = other_col
+                        break
+        return mapped_columns
+
     def _translate_columns(self, origin_df:pd.DataFrame):
         """
         Translate the dataframe from other search engines 
@@ -384,17 +409,11 @@ class PSMReaderBase(object):
         None
             Add information inplace into self._psm_df
         """
+        mapped_columns = self._find_mapped_columns(origin_df)
         self._psm_df = pd.DataFrame()
-        for col, map_col in self.column_mapping.items():
-            if isinstance(map_col, str):
-                if map_col in origin_df.columns:
-                    self._psm_df[col] = origin_df[map_col]
-            else:       
-                for other_col in map_col:
-                    if other_col in origin_df.columns:
-                        self._psm_df[col] = origin_df[other_col]
-                        break
-                    
+        for col, map_col in mapped_columns.items():
+            self._psm_df[col] = origin_df[map_col]
+               
         if (
             'scan_num' in self._psm_df.columns and 
             not 'spec_idx' in self._psm_df.columns