+++ docs/conf.py
@@ -23,7 +23,7 @@ project = 'alphabase'
 copyright = '2022, Mann Labs, MPIB'
 author = 'Mann Labs, MPIB'
 
-release = "0.5.0"
+release = "1.0.2"
 
 # -- General configuration ---------------------------------------------------