+++ misc/check_version.sh
@@ -1,6 +1,6 @@
 current_version=$(grep "__version__" ../alphabase/__init__.py | cut -f3 -d ' ' | sed 's/"//g')
 current_version_as_regex=$(echo $current_version | sed 's/\./\\./g')
-conda create -n version_check python=3.8 pip=20.1 -y
+conda create -n version_check python=3.9 -y
 conda activate version_check
 set +e
 already_on_pypi=$(pip install alphabase== 2>&1 | grep -c "$current_version_as_regex")