+++ release/one_click_windows_gui/create_installer_windows.sh
@@ -8,7 +8,7 @@ rm -rf dist
 rm -rf build
 
 # Creating a conda environment
-conda create -n alphabase_installer python=3.8 -y
+conda create -n alphabase_installer python=3.9 -y
 conda activate alphabase_installer
 
 # Creating the wheel
@@ -17,7 +17,7 @@ python setup.py sdist bdist_wheel
 # Setting up the local package
 cd release/one_click_windows_gui
 # Make sure you include the required extra packages and always use the stable or very-stable options!
-pip install "../../dist/alphabase-0.5.0-py3-none-any.whl[stable]"
+pip install "../../dist/alphabase-1.0.2-py3-none-any.whl[stable]"
 
 # Creating the stand-alone pyinstaller folder
 pip install pyinstaller