+++ release/pypi/prepare_pypi_wheel.sh
@@ -1,5 +1,5 @@
 cd ../..
-conda create -n alphabase_pypi_wheel python=3.8
+conda create -n alphabase_pypi_wheel python=3.9
 conda activate alphabase_pypi_wheel
 pip install twine
 rm -rf dist